Separate types and safety
